docker-polling-app/backend/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.2
Werkzeug==2.2.2
Flask-Cors==3.0.10
